refactor(improve): extract process-lock primitives to improve/locks.ts (D1 step 1)
Pure move of PROCESS_LOCK_DEFS + heldProcessLocks Set + acquire/release/reset
into improve/locks.ts (the Set is now module-private there). The exit backstop's
ownership-safe iteration becomes releaseHeldLocksIfOwned(pid); resetHeldProcessLocks
re-exported for the test seam. Drops the now-unused file-lock + ConfigError imports
from improve.ts. No behavior change (no RAII yet) — lock invariants + skip-if-locked
tests green; unit 5274/0.
